+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.5.0
 * expose `http` data as variable
+* allow using validators and track in definitions
 
 ## 0.4.0
 * add `flatten` prop to control processing multiple pages in one rule